admin_with_auth.py
#!/usr/bin/env python3
"""
Admin Server with Authentication
"""
import asyncio
import os
from fastapi import FastAPI, Depends, HTTPException, status
from fastapi.responses import FileResponse
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import JWTError, jwt
import uvicorn
from datetime import datetime, timedelta

# Set environment
os.environ["DATABASE_URL"] = "postgresql+asyncpg://postgres:password@localhost:5432/cricalgo"

app = FastAPI(title="CricAlgo Admin with Auth", version="1.0.0")

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Simple JWT secret for testing
JWT_SECRET = "test-secret-key-for-admin"
JWT_ALGORITHM = "HS256"

security = HTTPBearer()

def create_access_token(data: dict):
    """Create JWT token"""
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(hours=24)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, JWT_SECRET, algorithm=JWT_ALGORITHM)

def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    """Verify JWT token"""
    try:
        payload = jwt.decode(credentials.credentials, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        return payload
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication credentials"
        )

@app.get("/")
async def root():
    return {"message": "CricAlgo Admin with Auth", "status": "ok"}

@app.get("/admin")
async def admin_ui():
    return FileResponse("app/static/admin/index.html")

@app.get("/assets/{filename}")
async def admin_assets(filename: str):
    return FileResponse(f"app/static/admin/assets/{filename}")

# Login endpoint
@app.post("/api/v1/admin/login")
async def admin_login(credentials: dict):
    """Admin login"""
    username = credentials.get("username")
    password = credentials.get("password")
    
    # Simple hardcoded admin for testing
    if username == "admin" and password == "admin123":
        token = create_access_token({"username": "admin", "type": "admin"})
        return {
            "access_token": token,
            "token_type": "bearer",
            "user": {"username": "admin", "email": "admin@cricalgo.com"}
        }
    else:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid username or password"
        )

@app.get("/api/v1/admin/matches")
async def list_matches(token: dict = Depends(verify_token)):
    """List matches from database"""
    try:
        from app.db.session import get_db
        from app.models.match import Match
        from sqlalchemy import select
        
        async for db in get_db():
            result = await db.execute(select(Match).order_by(Match.created_at.desc()))
            matches = result.scalars().all()
            return [match.to_dict() for match in matches]
    except Exception as e:
        return {"error": f"Database error: {str(e)}", "matches": []}

@app.get("/api/v1/admin/deposits")
async def list_deposits(status: str = "pending", token: dict = Depends(verify_token)):
    """List deposits from database"""
    try:
        from app.db.session import get_db
        from app.models.transaction import Transaction
        from sqlalchemy import select
        
        async for db in get_db():
            if status == "pending":
                result = await db.execute(
                    select(Transaction).where(
                        Transaction.tx_type == "deposit",
                        Transaction.processed_at.is_(None)
                    ).order_by(Transaction.created_at.desc())
                )
            else:
                result = await db.execute(
                    select(Transaction).where(
                        Transaction.tx_type == "deposit",
                        Transaction.processed_at.isnot(None)
                    ).order_by(Transaction.created_at.desc())
                )
            transactions = result.scalars().all()
            return [tx.to_dict() for tx in transactions]
    except Exception as e:
        return {"error": f"Database error: {str(e)}", "deposits": []}

@app.post("/api/v1/admin/deposits/{tx_id}/approve")
async def approve_deposit(tx_id: str, payload: dict = None, token: dict = Depends(verify_token)):
    """Approve deposit"""
    try:
        from app.db.session import get_db
        from app.models.transaction import Transaction
        from sqlalchemy import select
        from datetime import datetime
        
        async for db in get_db():
            result = await db.execute(select(Transaction).where(Transaction.id == tx_id))
            tx = result.scalar_one_or_none()
            if not tx:
                return {"error": "Transaction not found"}
            
            tx.processed_at = datetime.utcnow()
            tx.status = "approved"
            await db.commit()
            
            return {"message": f"Deposit {tx_id} approved successfully!", "ok": True}
    except Exception as e:
        return {"error": f"Database error: {str(e)}"}

@app.post("/api/v1/admin/deposits/{tx_id}/reject")
async def reject_deposit(tx_id: str, payload: dict, token: dict = Depends(verify_token)):
    """Reject deposit"""
    try:
        from app.db.session import get_db
        from app.models.transaction import Transaction
        from sqlalchemy import select
        from datetime import datetime
        
        async for db in get_db():
            result = await db.execute(select(Transaction).where(Transaction.id == tx_id))
            tx = result.scalar_one_or_none()
            if not tx:
                return {"error": "Transaction not found"}
            
            tx.processed_at = datetime.utcnow()
            tx.status = "rejected"
            await db.commit()
            
            return {"message": f"Deposit {tx_id} rejected successfully!", "ok": True}
    except Exception as e:
        return {"error": f"Database error: {str(e)}"}

@app.get("/api/v1/admin/withdrawals")
async def list_withdrawals(status: str = "pending", token: dict = Depends(verify_token)):
    """List withdrawals from database"""
    try:
        from app.db.session import get_db
        from app.models.withdrawal import Withdrawal
        from sqlalchemy import select
        
        async for db in get_db():
            result = await db.execute(
                select(Withdrawal).where(Withdrawal.status == status)
                .order_by(Withdrawal.created_at.desc())
            )
            withdrawals = result.scalars().all()
            return [w.to_dict() for w in withdrawals]
    except Exception as e:
        return {"error": f"Database error: {str(e)}", "withdrawals": []}

@app.get("/api/v1/admin/audit")
async def get_audit(limit: int = 200, token: dict = Depends(verify_token)):
    """Get audit logs from database"""
    try:
        from app.db.session import get_db
        from app.models.audit_log import AuditLog
        from sqlalchemy import select
        
        async for db in get_db():
            result = await db.execute(
                select(AuditLog).order_by(AuditLog.created_at.desc()).limit(limit)
            )
            logs = result.scalars().all()
            return [log.to_dict() for log in logs]
    except Exception as e:
        return {"error": f"Database error: {str(e)}", "audit": []}

if __name__ == "__main__":
    print("🚀 Starting ADMIN SERVER with AUTHENTICATION...")
    print("📱 Admin UI: http://localhost:8000/admin")
    print("🔐 Admin Credentials:")
    print("   Username: admin")
    print("   Password: admin123")
    print("🗄️ Connected to PostgreSQL database")
    print("✅ This is a FULLY FUNCTIONAL admin panel with auth!")
    uvicorn.run(app, host="0.0.0.0", port=8000, log_level="info")
