requirements_no_k6.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis>=4.6.0

# Telegram Bot
aiogram==3.2.0

# Configuration
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
pytest-cov==4.1.0

# Celery
celery==5.3.4

# JWT Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Prometheus metrics
prometheus-client==0.19.0

# Sentry integration
sentry-sdk[fastapi]==1.38.0

# Linting and formatting
ruff==0.1.6
black==23.11.0

# Development
python-multipart==0.0.6

# Password hashing and 2FA
pyotp==2.9.0
